// src/modules/activity-report/domain/enums/logs.enum.tsx
export enum ActionLogs {
  POST_AUTH_TOKEN = 'post_oauth2_token',
  POST_PERM_PROFILES = 'post_perm_profiles',
  POST_PERM_PROFILES_ALL_IN_ONE = 'post_perm_profiles_all_in_one',
  POST_PERM_PROFILES_FEATURES = 'post_perm_profiles_features',
  DELETE_PERM_PROFILES_FEATURES = 'delete_perm_profiles_features',
  PATCH_PERM_PROFILES_STATUS = 'patch_perm_profiles_status',
  POST_OPERATIONS_USERS = 'post_operations_users',
  POST_USERS = 'post_users',
  PATCH_USERS_STATUS = 'patch_users_status',
  PUT_USERS_PASSWORD_RESETS = 'put_users_password_resets',
  PATCH_USERS = 'patch_users',
  PUT_USERS_PASSWORDS = 'put_users_passwords',
  POST_USERS_PASSWORDS = 'post_users_passwords',
  POST_USERS_PERM_PROFILES = 'post_users_perm_profiles',
  PUT_USERS_PERM_PROFILES_ALL_IN_ONE = 'put_users_perm_profiles_all_in_one',
  DELETE_USERS_PERM_PROFILES = 'delete_users_perm_profiles',
  POST_USERS_PERM_PROFILES_CONTRACTS = 'post_users_perm_profiles_contracts',
  DELETE_USERS_PERM_PROFILES_CONTRACTS = 'delete_users_perm_profiles_contracts',
  POST_CONTRACTS = 'post_contracts'
}